Converting classes in favor of functions + simplified generator API
