+++ generator/src/test/kotlin/com/patxi/poetimizely/generator/optimizely/OptimizelyTest.kt
@@ -1,4 +1,4 @@
-package com.patxi.poetimizely.optimizely
+package com.patxi.poetimizely.generator.optimizely
 
 import io.kotest.core.spec.style.BehaviorSpec
 import io.kotest.matchers.shouldBe
