+++ gradle-plugin/src/main/kotlin/com/patxi/poetimizely/gradle/plugin/PoetimizelyPlugin.kt
@@ -1,4 +1,4 @@
-package com.patxi.poetimizely.plugin
+package com.patxi.poetimizely.gradle.plugin
 
 import org.gradle.api.Plugin
 import org.gradle.api.Project
