+++ gradle-plugin/src/test/kotlin/com/patxi/poetimizely/gradle/plugin/BuildScriptTest.kt
@@ -1,4 +1,4 @@
-package com.patxi.poetimizely.plugin
+package com.patxi.poetimizely.gradle.plugin
 
 import io.kotest.core.spec.style.BehaviorSpec
 import io.kotest.matchers.shouldBe
